+++ ios/RNInstabug/InstabugReactBridge.m
@@ -382,6 +382,10 @@ RCT_EXPORT_METHOD(clearAllExperiments) {
     [Instabug clearAllExperiments];
 }
 
+RCT_EXPORT_METHOD(willRedirectToStore){
+    [Instabug willRedirectToAppStore];
+}
+
 - (NSDictionary *)constantsToExport {
     return ArgsRegistry.getAll;
 }